requirements.txt
Django==2.0.5    
django-neomodel
djangorestframework
django-rest-swagger
neo4j-driver==1.5.2
gunicorn